3-advanced-unit-testing-m3-structural-inspection-exercise-files/Shop/Shop.UnitTest/BasketTotalTests.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;
using Xunit;
using Ploeh.Samples.Shop;
using Moq;
using Xunit.Extensions;

namespace Ploeh.Samples.Shop.UnitTest
{
    public class BasketTotalTests
    {
        [Fact]
        public void SutIsBasketElement()
        {
            var sut = new BasketTotal();
            Assert.IsAssignableFrom<IBasketElement>(sut);
        }

        [Fact]
        public void AcceptReturnsCorrectResponse()
        {
            var r = new MockRepository(MockBehavior.Default)
            {
                DefaultValue = DefaultValue.Mock
            };
            var expected = r.Create<IBasketVisitor>().Object;
            var sut = new BasketTotal();

            var visitorStub = r.Create<IBasketVisitor>();
            visitorStub.Setup(v => v.Visit(sut)).Returns(expected);
            IBasketVisitor actual = sut.Accept(visitorStub.Object);

            Assert.Same(expected, actual);
        }

        [Fact]
        public void InitialTotalIsCorrect()
        {
            var sut = new BasketTotal();
            decimal actual = sut.Total;
            Assert.Equal(0, actual);
        }

        [Theory]
        [InlineData(1)]
        [InlineData(2)]
        public void TotalIsCorrect(int expected)
        {
            var sut = new BasketTotal(expected);
            var actual = sut.Total;
            Assert.Equal(expected, actual);
        }
    }
}
